Move API patching to ifcopenshell.api
